first commit
Create an ArchiveViewController, which displays posts using a CollectionView with a nav bar at the top. The view controller stores content as ContentMetadata, and shows a loading indicator when no data is available. Data is fetched using ContentService. The layout of the collection view is defined by class constants.

Create a ContentCell class which represents a cell of content in a CollectionView. display image, date, comments, and likes data in a square cell. On long press, displays a share/info/delete pop up.

Create a ContentService singleton class which loads data for the ArchiveViewController. For now, we simulate an API call by forcing a delay and returning mock data.

the ContentMetadata struct defines the shape of data we expect.

Todo items:
- menu button doesn't have functionality
- real API call integration
- functionality to open a separate view to see a single cell of content
- multiple tabs of content